Removed Padim from init
