+++ anomalib/models/padim/model.py
@@ -26,7 +26,6 @@ import torch.nn.functional as F
 import torchvision
 from kornia import gaussian_blur2d
 from omegaconf import DictConfig, ListConfig
-from pytorch_lightning.utilities.cli import MODEL_REGISTRY
 from torch import Tensor, nn
 
 from anomalib.models.components import (
@@ -285,92 +284,6 @@ class PadimLightning(AnomalyModule):
             layers=hparams.model.layers,
             input_size=hparams.model.input_size,
             backbone=hparams.model.backbone,
-        ).eval()
-
-        self.stats: List[Tensor] = []
-        self.automatic_optimization = False
-        self.embeddings: List[Tensor] = []
-
-    @staticmethod
-    def configure_optimizers():
-        """PADIM doesn't require optimization, therefore returns no optimizers."""
-        return None
-
-    def training_step(self, batch, _batch_idx):  # pylint: disable=arguments-differ
-        """Training Step of PADIM. For each batch, hierarchical features are extracted from the CNN.
-
-        Args:
-            batch (Dict[str, Any]): Batch containing image filename, image, label and mask
-            _batch_idx: Index of the batch.
-
-        Returns:
-            Hierarchical feature map
-        """
-        self.model.feature_extractor.eval()
-        embedding = self.model(batch["image"])
-
-        # NOTE: `self.embedding` appends each batch embedding to
-        #   store the training set embedding. We manually append these
-        #   values mainly due to the new order of hooks introduced after PL v1.4.0
-        #   https://github.com/PyTorchLightning/pytorch-lightning/pull/7357
-        self.embeddings.append(embedding.cpu())
-
-    def on_validation_start(self) -> None:
-        """Fit a Gaussian to the embedding collected from the training set."""
-        # NOTE: Previous anomalib versions fit Gaussian at the end of the epoch.
-        #   This is not possible anymore with PyTorch Lightning v1.4.0 since validation
-        #   is run within train epoch.
-        embeddings = torch.vstack(self.embeddings)
-        self.stats = self.model.gaussian.fit(embeddings)
-
-    def validation_step(self, batch, _):  # pylint: disable=arguments-differ
-        """Validation Step of PADIM.
-
-        Similar to the training step, hierarchical features are extracted from the CNN for each batch.
-
-        Args:
-            batch: Input batch
-            _: Index of the batch.
-
-        Returns:
-            Dictionary containing images, features, true labels and masks.
-            These are required in `validation_epoch_end` for feature concatenation.
-        """
-
-        batch["anomaly_maps"] = self.model(batch["image"])
-        return batch
-
-
-@MODEL_REGISTRY
-class Padim(AnomalyModule):
-    """PaDiM: a Patch Distribution Modeling Framework for Anomaly Detection and Localization.
-
-    Args:
-        task (str): Task type (classification | segmentation)
-        adaptive_threshold (bool): Boolean to automatically choose adaptive threshold
-        default_image_threshold (float): Manual default image threshold
-        default_pixel_threshold (float): Manaul default pixel threshold
-        layers (List[str]): Layers to extract features from the backbone CNN
-        input_size (Tuple[int, int]): Size of the model input.
-        backbone (str): Backbone CNN network
-    """
-
-    def __init__(
-        self,
-        task: str,
-        adaptive_threshold: bool,
-        default_image_threshold: float,
-        default_pixel_threshold: float,
-        layers: List[str],
-        input_size: Tuple[int, int],
-        backbone: str,
-    ):
-        super().__init__(task, adaptive_threshold, default_image_threshold, default_pixel_threshold)
-        self.layers = layers
-        self.model: PadimModel = PadimModel(
-            layers=layers,
-            input_size=input_size,
-            backbone=backbone,
         ).eval()
 
         self.stats: List[Tensor] = []
