+++ anomalib/models/patchcore/model.py
@@ -109,28 +109,21 @@ class PatchcoreModel(DynamicBufferModule, nn.Module):
 
     def __init__(
         self,
-        layers: List[str],
         input_size: Tuple[int, int],
+        layers: List[str],
         backbone: str = "wide_resnet50_2",
-        apply_tiling: bool = False,
-        tile_size: Optional[Tuple[int, int]] = None,
-        tile_stride: Optional[int] = None,
     ) -> None:
         super().__init__()
 
+        self.tiler: Optional[Tiler] = None
+
+        self.input_size = input_size
         self.backbone = getattr(torchvision.models, backbone)
         self.layers = layers
-        self.input_size = input_size
-        self.apply_tiling = apply_tiling
 
         self.feature_extractor = FeatureExtractor(backbone=self.backbone(pretrained=True), layers=self.layers)
         self.feature_pooler = torch.nn.AvgPool2d(3, 1, 1)
         self.anomaly_map_generator = AnomalyMapGenerator(input_size=input_size)
-
-        if apply_tiling:
-            assert tile_size is not None
-            assert tile_stride is not None
-            self.tiler = Tiler(tile_size, tile_stride)
 
         self.register_buffer("memory_bank", torch.Tensor())
         self.memory_bank: torch.Tensor
@@ -150,7 +143,7 @@ class PatchcoreModel(DynamicBufferModule, nn.Module):
             Union[torch.Tensor, Tuple[torch.Tensor, torch.Tensor]]: Embedding for training,
                 anomaly map and anomaly score for testing.
         """
-        if self.apply_tiling:
+        if self.tiler:
             input_tensor = self.tiler.tile(input_tensor)
 
         with torch.no_grad():
@@ -159,7 +152,7 @@ class PatchcoreModel(DynamicBufferModule, nn.Module):
         features = {layer: self.feature_pooler(feature) for layer, feature in features.items()}
         embedding = self.generate_embedding(features)
 
-        if self.apply_tiling:
+        if self.tiler:
             embedding = self.tiler.untile(embedding)
 
         embedding = self.reshape_embedding(embedding)
@@ -252,16 +245,22 @@ class PatchcoreLightning(AnomalyModule):
         apply_tiling (bool, optional): Apply tiling. Defaults to False.
     """
 
-    def __init__(self, hparams) -> None:
-        super().__init__(hparams)
+    def __init__(
+        self,
+        task: str,
+        input_size: Tuple[int, int],
+        layers: List[str],
+        backbone: str,
+        adaptive_threshold: bool,
+        default_image_threshold: float,
+        default_pixel_threshold: float,
+    ) -> None:
+        super().__init__(task, adaptive_threshold, default_image_threshold, default_pixel_threshold)
 
         self.model: PatchcoreModel = PatchcoreModel(
-            layers=hparams.model.layers,
-            input_size=hparams.model.input_size,
-            tile_size=hparams.dataset.tiling.tile_size,
-            tile_stride=hparams.dataset.tiling.stride,
-            backbone=hparams.model.backbone,
-            apply_tiling=hparams.dataset.tiling.apply,
+            input_size=input_size,
+            layers=layers,
+            backbone=backbone,
         )
         self.automatic_optimization = False
         self.embeddings: List[Tensor] = []
